ENG-7284 - Refactor pt.2 - Put all extra utils for RN/Android into one folder
