--- NeuroID/src/android/java/com/neuroid/tracker/callbacks/NIDActivityCallbackUtil.kt
@@ -1,6 +1,0 @@
-package com.neuroid.tracker.callbacks
-
-// Native Android needs to register the activities
-internal fun registrationHelpers(r:Runnable){
-    r.run()
-}
