--- NeuroID/src/android/java/com/neuroid/tracker/events/NIDIdentifyAllViews.kt
@@ -1,7 +1,0 @@
-package com.neuroid.tracker.events
-
-import android.view.View
-
-fun verifyComponentType(view: View): ComponentValuesResult {
-    return isCommonAndroidComponent(view)
-}
