--- NeuroID/src/android/java/com/neuroid/tracker/events/NIDRegisterEventsHelper.kt
@@ -1,6 +1,0 @@
-package com.neuroid.tracker.events
-
-// run the function immediately in native android
-internal fun handleIdentifyAllViews(r:Runnable){
-    r.run()
-}
