--- NeuroID/src/reactNative/java/com/neuroid/tracker/callbacks/NIDActivityCallbackUtil.kt
@@ -1,6 +1,0 @@
-package com.neuroid.tracker.callbacks
-
-// In RN users call registerPageTargets so we don't
-//    want to run this code and double register
-internal fun registrationHelpers(r:Runnable){
-}
