--- NeuroID/src/reactNative/java/com/neuroid/tracker/events/NIDRegisterEventsHelper.kt
@@ -1,10 +1,0 @@
-package com.neuroid.tracker.events
-
-import android.os.Looper
-
-// run the function after a 300ms delay in react native
-internal fun handleIdentifyAllViews(r:Runnable){
-    android.os.Handler(Looper.getMainLooper()).postDelayed({
-        r.run()
-    }, 300)
-}
