+++ NeuroID/src/reactNative/java/com/neuroid/tracker/utils/ExtraUtils.kt
@@ -1,14 +1,23 @@
-package com.neuroid.tracker.events
+package com.neuroid.tracker.utils
 
+import android.os.Looper
 import android.view.View
-import com.neuroid.tracker.extensions.getSHA256withSalt
-import com.neuroid.tracker.extensions.getIdOrTag
-
 import androidx.core.view.children
 import com.facebook.react.views.image.ReactImageView
 import com.facebook.react.views.text.ReactTextView
 import com.facebook.react.views.textinput.ReactEditText
 import com.facebook.react.views.view.ReactViewGroup
+
+import com.neuroid.tracker.events.ComponentValuesResult
+import com.neuroid.tracker.events.detectBasicAndroidViewType
+import com.neuroid.tracker.events.isCommonAndroidComponent
+import com.neuroid.tracker.extensions.getSHA256withSalt
+import com.neuroid.tracker.extensions.getIdOrTag
+
+// In RN users call registerPageTargets so we don't
+//    want to run this code and double register
+internal fun registrationHelpers(r:Runnable){
+}
 
 fun isCommonReactNativeComponent(view: View): ComponentValuesResult {
     // vars that can change based on the element
@@ -50,8 +59,7 @@ fun isCommonReactNativeComponent(view: View): ComponentValuesResult {
     )
 }
 
-
-fun verifyComponentType(view: View):ComponentValuesResult{
+fun verifyComponentType(view: View): ComponentValuesResult {
     var (idName, et, v, metaData) = isCommonAndroidComponent(view)
 
     // check if its empty, then check if its a RN element
@@ -70,4 +78,37 @@ fun verifyComponentType(view: View):ComponentValuesResult{
         v,
         metaData
     )
+}
+
+// run the function after a 300ms delay in react native
+internal fun handleIdentifyAllViews(r:Runnable){
+    android.os.Handler(Looper.getMainLooper()).postDelayed({
+        r.run()
+    }, 300)
+}
+
+
+internal fun detectViewType(currentView:View?):Int {
+    var typeOfView = when (currentView) {
+        is ReactEditText
+        -> 1
+        is ReactViewGroup -> {
+            if (currentView.isFocusable) {
+                2
+            } else {
+                0
+            }
+        }
+        else -> 0
+    }
+
+    if (typeOfView == 0) {
+        typeOfView  = detectBasicAndroidViewType(currentView)
+    }
+
+    return typeOfView
+}
+
+internal fun getEtnSenderName(currentView:View?): String{
+    return currentView?.getIdOrTag() ?: "main_view"
 }
